Reformat db section
Try not to break up sections into individual sentences unless it really is an ordered list of things or a bullet point list of related items. I do this a lot myself where I want to break it down to each single idea, but most of the time it should just be a 3-5 sentence paragraph.