Create code blocks for terminal commands
This part is really important because people are going to trying to copy and paste these commands to get the project to work. You need to be very explicit about what the exact commands are to the point where you can take each command and copy and paste it from the markdown into a terminal and have the whole project run.

You could keep it the way you had it and just make sure that you carefully mark all the commands with ``, but I find this introduces way more room for error.

Also I haven't actually run the project, but it looks like you need to actually provide the commands to navigate out of the server folder and into the frontend. You include that instruction in words, but if someone just took the next command it would simply run `yarn install` again in the same place they just were. So I think you'd want to do something like:

```bash
cd ..
cd frontend
yarn install
```